AudioPolicy: fix stopInput
Given the convention for audio device identification since JB MR1 code
drop, the bit 31 of the bit field is reserved and when set indicates
an input device. So 0x0 (AUDIO_DEVICE_NONE) is identified as an output
audio device. Thus the stopInput function does not stop the audio input
stream.

In order to reroute the audio input stream as expected, this patch
modifies the stopInput function in order to use the
AUDIO_DEVICE_BIT_IN (0x80000000) value.

Change-Id: I0ee2f863cbf9446e9cf0e6d388b3ac2670faafee
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Signed-off-by: Jin Wei <wei.a.jin@intel.com>